% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_clean.R
\name{ad_clean}
\alias{ad_clean}
\title{A function to retrieve the anomalies using rds.}
\usage{
ad_clean(log_data, document, recorded_date, logs)
}
\arguments{
\item{log_data}{A dataset of logs aggregated by publications and per days.}
}
\description{
A function to retrieve the anomalies using rds.
}
\examples{
 ad_clean(log_data)
}
