% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_model.R
\name{um_model}
\alias{um_model}
\title{A function to retrieve the anomalies using rds.}
\usage{
um_model(log_data, knots = 3, min_observations = 7,
  min_anomaly_ratio = 5)
}
\arguments{
\item{log_data}{A dataset of logs aggregated by publications and per days.}

\item{knots}{The number of knots fitted to the knots. By default 3. A larger value may better fit log data with a long/complicated history, at the expense of being to overfitted to anomalous episode.}

\item{min_observations}{The minimal number of observations to derive the model. By default 6 with 3 knots.}

\item{min_anomaly_ratio}{The threshold of anomaly rate (by default 5 but can be any number between 0 and 100)}
}
\description{
A function to retrieve the anomalies using rds.
}
\examples{
 um_model(log_data, knots = 3, min_observations = 7, min_anomaly_ratio = 5)
}
