% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_rcs_anomalies.R
\name{um_rcs_anomalies}
\alias{um_rcs_anomalies}
\title{A function to retrieve the anomalous ratio using the contrast with the predicted value in the rcs model.
The function is preferably used within do or map.}
\usage{
um_rcs_anomalies(fit, count_log, time_sequency, knots = 3)
}
\arguments{
\item{fit}{An rcs model already fitted to the aggregated logs.}

\item{count_log}{A vector of number of aggregated logs}

\item{time_sequency}{A vector containing a time sequence (usually days) on which the aggregation has been performed.}

\item{knots}{The number of knots fitted to the knots}
}
\value{
A vector of confidence interval in regards to the model.
}
\description{
A function to retrieve the anomalous ratio using the contrast with the predicted value in the rcs model.
The function is preferably used within do or map.
}
\examples{
 um_rcs_anomalies(fit, count_log, time_sequency, knots)
}
