% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_rcs_group.R
\name{um_rcs_group}
\alias{um_rcs_group}
\title{A function to apply an RCS model to log data.}
\usage{
um_rcs_group(time_sequency, count_log, knots = 3)
}
\arguments{
\item{time_sequency}{A vector containing the list of day since the start of the recording.}

\item{count_log}{A vector of number of visitor per day.}

\item{knots}{The number of knots fitted to the knots. By default 3. A larger value may better fit log data with a long/complicated history, at the expense of being to overfitted to anomalous episode.}
}
\value{
The results of the SVM model.
um_rcs_group(time_sequency, count_log, knots=3)
}
\description{
A function to apply an RCS model to log data.
}
