% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_clean.R
\name{um_clean}
\alias{um_clean}
\title{A function to retrieve the anomalies using rds.}
\usage{
um_clean(log_data, document, time_variable, count_log)
}
\arguments{
\item{log_data}{A dataset of logs aggregated by publications and per days.}

\item{document}{The exact name of the document column.}

\item{count_log}{The exact name of the aggregated logs column.}

\item{time_sequence}{The exact name of the date of records column.}
}
\description{
A function to retrieve the anomalies using rds.
}
\examples{
 um_clean(log_data, document, recorded_date, logs)
}
